Almost complete getting access denied error when launching apps currently
