+++ appLauncher.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30002.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "appLauncher", "appLauncher\appLauncher.csproj", "{825249DF-2022-4222-96A1-69A32AA4E16C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Concurrent", "Concurrent\Concurrent.csproj", "{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,26 @@ Global
 		{825249DF-2022-4222-96A1-69A32AA4E16C}.Release|x86.ActiveCfg = Release|x86
 		{825249DF-2022-4222-96A1-69A32AA4E16C}.Release|x86.Build.0 = Release|x86
 		{825249DF-2022-4222-96A1-69A32AA4E16C}.Release|x86.Deploy.0 = Release|x86
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|ARM.ActiveCfg = Debug|ARM
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|ARM.Build.0 = Debug|ARM
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|ARM64.Build.0 = Debug|ARM64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|x64.ActiveCfg = Debug|x64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|x64.Build.0 = Debug|x64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|x86.ActiveCfg = Debug|x86
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Debug|x86.Build.0 = Debug|x86
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|ARM.ActiveCfg = Release|ARM
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|ARM.Build.0 = Release|ARM
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|ARM64.ActiveCfg = Release|ARM64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|ARM64.Build.0 = Release|ARM64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|x64.ActiveCfg = Release|x64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|x64.Build.0 = Release|x64
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|x86.ActiveCfg = Release|x86
+		{280ED59E-E4B3-4860-86AF-E2EF6F947DBF}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
