+++ appLauncher/Core/Pages/MainPage.xaml
@@ -1,6 +1,6 @@
 ﻿<!--View of MainPage. Here is where the apps are displayed-->
 <Page
-    x:Class="appLauncher.MainPage"
+    x:Class="appLauncher.mobile.Core.Pages.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
