+++ appLauncher/Core/Pages/SearchPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="appLauncher.SearchPage"
+    x:Class="appLauncher.mobile.Core.Pages.SearchPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:appLauncher"
