+++ appLauncher/Core/Pages/settings.xaml
@@ -1,5 +1,5 @@
 ﻿<!--The view of the settings page.-->
-    <Page x:Class="appLauncher.Pages.settings"
+    <Page x:Class="appLauncher.mobile.Core.Pages.settings"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:local="using:appLauncher.Pages"
