+++ appLauncher/Core/Pages/splashScreen.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="appLauncher.splashScreen"
+    x:Class="appLauncher.mobile.Core.Pages.splashScreen"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:appLauncher"
