+++ appLauncher/Core/Pages/splashScreen.xaml.cs
@@ -1,26 +1,21 @@
 ﻿using appLauncher.mobile.Core.Helpers;
+
 using Microsoft.Toolkit.Uwp.UI.Animations;
+
 using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
 using System.Threading.Tasks;
+using System.Runtime;
 using Windows.ApplicationModel.Activation;
 using Windows.Foundation;
-using Windows.Foundation.Collections;
 using Windows.UI.Core;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace appLauncher
+namespace appLauncher.mobile.Core.Pages
+
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
@@ -55,7 +50,7 @@ namespace appLauncher
             }
 
             // Create a Frame to act as the navigation context
-            
+
 
         }
 
@@ -67,7 +62,7 @@ namespace appLauncher
         protected override void OnNavigatedTo(NavigationEventArgs e)
         {
             base.OnNavigatedTo(e);
-            
+
         }
 
         private async void DismissedEventHandler(SplashScreen sender, object args)
@@ -86,18 +81,18 @@ namespace appLauncher
 
             dismissed = true;
 
-            
-                await Windows.ApplicationModel.Core.CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, async () =>
-                {
-                    while (appsLoaded == false)
-                    {
-                        await theImage.Scale(0.9f, 0.9f, (float)theImage.ActualWidth / 2, (float)theImage.ActualHeight / 2, 1000, 0, EasingType.Linear).StartAsync();
-                        await theImage.Scale(1f, 1f, (float)theImage.ActualWidth / 2, (float)theImage.ActualHeight / 2, 1000, 0, EasingType.Linear).StartAsync();
 
-                        
-                    }
-                    
-                });
+            await Windows.ApplicationModel.Core.CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, async () =>
+            {
+                while (appsLoaded == false)
+                {
+                    await theImage.Scale(0.9f, 0.9f, (float)theImage.ActualWidth / 2, (float)theImage.ActualHeight / 2, 1000, 0, EasingType.Linear).StartAsync();
+                    await theImage.Scale(1f, 1f, (float)theImage.ActualWidth / 2, (float)theImage.ActualHeight / 2, 1000, 0, EasingType.Linear).StartAsync();
+
+
+                }
+
+            });
 
 
             //await Task.Run(() => finalAppItem.getApps());
@@ -106,16 +101,16 @@ namespace appLauncher
             await packageHelper.getAllAppsAsync();
             //await GlobalVariables.LoadCollectionAsync();
             await Task.Delay(1500);
-        
+
 
 
             // Complete app setup operations here...
 
         }
 
-        
 
-       
+
+
 
         public async void DismissExtendedSplash()
         {
@@ -131,12 +126,12 @@ namespace appLauncher
                 var imagePosX = visualStuff.X;
 
                 var imageXToTravelTo = width - imagePosX;
-                
-              
+
+
 
                 await theImage.Offset(-100, 100).StartAsync();
-                var anim = theImage.Offset((float)width/2, (float)-height / 2, 100, 0, EasingType.Cubic).Fade(0,50,50);
-                
+                var anim = theImage.Offset((float)width / 2, (float)-height / 2, 100, 0, EasingType.Cubic).Fade(0, 50, 50);
+
 
                 anim.Completed += Anim_Completed;
                 await anim.StartAsync();
@@ -145,7 +140,7 @@ namespace appLauncher
             });
         }
 
-       
+
         private void Anim_Completed(object sender, AnimationSetCompletedEventArgs e)
         {
