+++ appLauncher/Package.appxmanifest
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Package xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest" xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10" xmlns:iot="http://schemas.microsoft.com/appx/manifest/iot/windows10" xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities" IgnorableNamespaces="uap mp iot rescap">
-  <Identity Name="c7a9e123-47c7-47c2-9040-99dcd4ac7351" Publisher="CN=colin" Version="1.2.4.0" />
+  <Identity Name="c7a9e123-47c7-47c2-9040-99dcd4ac7351" Publisher="CN=Isaac" Version="1.2.4.0" />
   <mp:PhoneIdentity PhoneProductId="c7a9e123-47c7-47c2-9040-99dcd4ac7351" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
   <Properties>
     <DisplayName>appLauncher</DisplayName>
@@ -23,9 +23,10 @@
             <uap:ShowOn Tile="square310x310Logo" />
           </uap:ShowNameOnTiles>
         </uap:DefaultTile>
-        <uap:SplashScreen Image="Assets\SplashScreen.png" />
+        <uap:SplashScreen Image="Core\Assets\SplashScreen.png" />
         <uap:InitialRotationPreference>
           <uap:Rotation Preference="portrait" />
+          <uap:Rotation Preference="landscape"/>
         </uap:InitialRotationPreference>
       </uap:VisualElements>
     </Application>
